+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 **Fixed**:
 
 - **decidim-blog**: Add `params[:id]` when editing/deleting a post from admin site [\#3360](https://github.com/decidim/decidim/pull/3360)
+- **decidim-admin**: Fixes the validation uniqueness name of area, scoped with organization and area_type [\#3350](https://github.com/decidim/decidim/pull/3350)
+
 ## [0.11.0.pre](https://github.com/decidim/decidim/tree/v0.11.0)
 
 **Upgrade notes**: